refactor: publishLocalBranch
