+++ electron/app/git/ipc.ts
@@ -2,7 +2,7 @@ import {ipcMain} from 'electron';
 
 import type {FileMapType} from '@shared/models/appState';
 
-import {getChangedFiles, getCommitResources, getGitRepoInfo, initGitRepo, publishLocalBranch, setRemote} from './git';
+import {getChangedFiles, getCommitResources, getGitRepoInfo, initGitRepo, setRemote} from './git';
 
 ipcMain.on('git.getGitRepoInfo', async (event, localPath: string) => {
   try {
@@ -28,15 +28,6 @@ ipcMain.on('git.getChangedFiles', async (event, payload: {localPath: string; fil
     event.sender.send('git.getChangedFiles.result', result);
   } catch (e: any) {
     event.sender.send('git.getChangedFiles.result', {error: e.message});
-  }
-});
-
-ipcMain.on('git.publishLocalBranch', async (event, payload: {localPath: string; branchName: string}) => {
-  try {
-    const result = await publishLocalBranch(payload.localPath, payload.branchName);
-    event.sender.send('git.publishLocalBranch.result', result);
-  } catch (e: any) {
-    event.sender.send('git.publishLocalBranch.result', {error: e.message});
   }
 });
