+++ electron/app/services/git/handlers/cloneGitRepo.ts
@@ -1,7 +1,7 @@
 import {promises as fs} from 'fs';
 import simpleGit from 'simple-git';
 
-import {GitCloneRepoParams} from '@shared/ipc';
+import {GitCloneRepoParams} from '@shared/ipc/git';
 
 export async function cloneGitRepo({localPath, repoPath}: GitCloneRepoParams): Promise<void> {
   try {
