+++ electron/app/services/git/handlers/index.ts
@@ -9,6 +9,7 @@ export * from './getBranchCommits';
 export * from './getGitRemotePath';
 export * from './isFolderGitRepo';
 export * from './isGitInstalled';
+export * from './publishLocalBranch';
 export * from './pullChanges';
 export * from './pushChanges';
 export * from './stageChangedFiles';
