+++ electron/app/services/git/handlers/pullChanges.ts
@@ -1,6 +1,6 @@
 import simpleGit from 'simple-git';
 
-import {GitPathParams} from '@shared/ipc';
+import {GitPathParams} from '@shared/ipc/git';
 
 export async function pullChanges({path}: GitPathParams): Promise<void> {
   const git = simpleGit({baseDir: path});
