+++ src/components/organisms/GitPane/BottomActions.tsx
@@ -8,11 +8,10 @@ import {TOOLTIP_DELAY} from '@constants/constants';
 import {GitCommitDisabledTooltip, GitCommitEnabledTooltip} from '@constants/tooltips';
 
 import {addGitBranch, setGitLoading} from '@redux/git';
-import {fetchRepo, pullChanges, pushChanges} from '@redux/git/service';
+import {fetchRepo, publishLocalBranch, pullChanges, pushChanges} from '@redux/git/service';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setAlert} from '@redux/reducers/alert';
 
-import {promiseFromIpcRenderer} from '@utils/promises';
 import {showGitErrorModal} from '@utils/terminal';
 
 import {AlertEnum} from '@shared/models/alert';
@@ -26,7 +25,7 @@ const BottomActions: React.FC = () => {
   const currentBranch = useAppSelector(state => state.git.repo?.currentBranch);
   const gitLoading = useAppSelector(state => state.git.loading);
   const gitRepo = useAppSelector(state => state.git.repo);
-  const selectedProjectRootFolder = useAppSelector(state => state.config.selectedProjectRootFolder);
+  const selectedProjectRootFolder = useAppSelector(state => state.config.selectedProjectRootFolder) || '';
 
   const [showCommitModal, setShowCommitModal] = useState(false);
 
@@ -91,20 +90,16 @@ const BottomActions: React.FC = () => {
 
     dispatch(setGitLoading(true));
 
-    const result = await promiseFromIpcRenderer('git.publishLocalBranch', 'git.publishLocalBranch.result', {
-      localPath: selectedProjectRootFolder,
-      branchName: currentBranch || 'main',
-    });
-
-    if (result.error) {
+    try {
+      await publishLocalBranch({localPath: selectedProjectRootFolder, branchName: currentBranch || 'main'});
+      dispatch(setAlert({title: 'Branch published successfully', message: '', type: AlertEnum.Success}));
+    } catch (e) {
       showGitErrorModal('Publishing local branch failed', `git push -u origin ${currentBranch || 'main'}`, dispatch);
       setGitLoading(false);
       return;
     }
 
     dispatch(addGitBranch(`origin/${currentBranch}`));
-    dispatch(setAlert({title: 'Branch published successfully', message: '', type: AlertEnum.Success}));
-
     dispatch(setGitLoading(false));
   }, [currentBranch, dispatch, gitRepo, selectedProjectRootFolder]);
